python/azure-ai-projects/requirements.txt
opentelemetry-sdk==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
azure-ai-projects==1.0.0b11
azure-identity==1.19.0
openai==1.85.0
opentelemetry-instrumentation-openai-v2==2.1b0
azure-ai-inference[opentelemetry]==1.0.0b9